Super-basic syntax highlighting
