+++ Cargo.toml
@@ -6,6 +6,7 @@ edition = "2024"
 [dependencies]
 clap = { version = "4.5.45", features = ["derive"] }
 ctrlc = "3.4.7"
+nu-ansi-term = "0.50.1"
 reedline = "0.41.0"
 thiserror = "2.0.16"
