+++ src/repl.rs
@@ -1,118 +1,11 @@
 use std::env;
 use std::io::{self, IsTerminal, Write};
-use clap::Args;
-use reedline::{Signal, DefaultPrompt, DefaultPromptSegment, HistoryItem};
+use reedline::{Signal, DefaultPrompt, DefaultPromptSegment, HistoryItem, Highlighter, StyledText};
+use nu_ansi_term::{Color, Style};
 
 use crate::{cli_util, BrainfuckReader};
 
-#[derive(Args, Debug)]
-#[command(disable_help_flag = true)]
-pub struct ReplArgs {
-    /// Force non-interactive bare mode
-    #[arg(long = "bare", conflicts_with = "editor")]
-    pub bare: bool,
-
-    /// Force interactive mode (errors if stdin is not a TTY)
-    #[arg(long = "editor", conflicts_with = "bare")]
-    pub editor: bool,
-
-    /// Show this help
-    #[arg(short = 'h', long = "help", action = clap::ArgAction::SetTrue)]
-    help: bool,
-}
-
-
-// Public entry point for the REPL from main.rs
-pub fn run(program: &str, help: bool, mode_flag: ModeFlagOverride) -> i32 {
-    if help {
-        usage_and_exit(program, 0);
-    }
-
-    // Determine mode: flags -> env -> auto-detect via is_terminal()
-    let mode = match select_mode(mode_flag) {
-        Ok(m) => m,
-        Err(msg) => {
-            eprintln!("{program}: {msg}");
-            let _ = io::stderr().flush();
-            return 1;
-        }
-    };
-
-    // Install SIGINT (ctrl+c) handler to flush and exit(0) immediately
-    if let Err(e) = ctrlc::set_handler(|| {
-        let _ = io::stdout().flush();
-        let _ = io::stderr().flush();
-        std::process::exit(0);
-    }) {
-        eprintln!("{program}: failed to set ctrl+c handler: {e}");
-        let _ = io::stderr().flush();
-        return 1;
-    }
-
-    match mode {
-        ReplMode::Editor => {
-            // Print banners/prompts only if stderr is a TTY
-            if io::stderr().is_terminal() {
-                eprintln!("Brainfuck REPL (interactive editor mode)");
-                eprintln!("Ctrl+d/Ctrl+z Enter (Windows) executes the current buffer. Press ctrl+c to exit");
-                let _ = io::stderr().flush();
-            }
-
-            if let Err(e) = repl_loop() {
-                eprintln!("{program}: REPL error: {e}");
-                let _ = io::stderr().flush();
-                return 1;
-            }
-
-            0
-        }
-        ReplMode::Bare => {
-            // Bare mode: read stdin until EOF, execute once, exit 0
-            match execute_bare_once() {
-                Ok(_) => 0,
-                Err(e) => {
-                    eprintln!("{program}: REPL error: {e}");
-                    let _ = io::stderr().flush();
-                    1
-                }
-            }
-        }
-    }
-}
-
-fn usage_and_exit(program: &str, code: i32) -> ! {
-    eprintln!(
-        r#"Usage:
-  {0} repl   # Start a Brainfuck REPL (read-eval-print loop)
-
-Options:
-  --help,   -h        Show this help
-  --bare              Force non-interactive bare mode
-  --editor            Force interactive editor mode (errors if stdin is not a TTY)
-
-Description:
-  Starts a REPL where you can enter Brainfuck code and execute it live.
-
-Notes:
-    - Ctrl+d executes the current buffer on *nix/macOS.
-    - Ctrl+z and Enter will execute the current buffer on Windows.
-    - Ctrl+c exits the REPL immediately.
-    - The REPL will print a newline after each execution for readability.
-    - Each execution starts with a fresh memory and pointer.
-    - The REPL will exit after a single execution if the environment variable `BF_REPL_ONCE` is set to `1`.
-    - Mode selection:
-        * Flags: --bare|--editor override environment and auto-detection.
-        * Env: BF_REPL_MODE=bare|editor overrides auto-detection (flags, when preset, will override env).
-        * Auto-detect: if stdin is a TTY, starts in interactive editor mode; otherwise, bare mode.
-        * Prompts/banners suppressed if stderr is not a TTY.
-"#,
-        program
-    );
-    let _ = io::stderr().flush();
-    std::process::exit(code);
-}
-
-fn repl_loop() -> io::Result<()> {
+pub fn repl_loop() -> io::Result<()> {
     // Initialize interactive line editor
     let mut editor = init_line_editor()?;
 
@@ -165,6 +58,7 @@ fn init_line_editor() -> io::Result<reedline::Reedline> {
     let history = reedline::FileBackedHistory::new(1_000).unwrap();
 
     let editor = Reedline::create()
+        .with_highlighter(Box::new(BrainfuckHighlighter::default()))
         .with_history(Box::new(history))
         .with_edit_mode(Box::new(Emacs::new(keybindings)));
 
@@ -250,7 +144,7 @@ fn execute_bf_buffer(buffer: String) {
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ReplMode {
+pub enum ReplMode {
     Bare,
     Editor,
 }
@@ -262,7 +156,7 @@ pub enum ModeFlagOverride {
     Editor,
 }
 
-fn select_mode(flag: ModeFlagOverride) -> Result<ReplMode, String> {
+pub fn select_mode(flag: ModeFlagOverride) -> Result<ReplMode, String> {
     // Flag override
     match flag {
         ModeFlagOverride::Bare => return Ok(ReplMode::Bare),
@@ -298,7 +192,7 @@ fn select_mode(flag: ModeFlagOverride) -> Result<ReplMode, String> {
     }
 }
 
-fn execute_bare_once() -> io::Result<()> {
+pub fn execute_bare_once() -> io::Result<()> {
     let mut locked = io::BufReader::new(io::stdin().lock());
     let submission = read_submission(&mut locked);
     if let Some(s) = submission {
@@ -311,6 +205,50 @@ fn execute_bare_once() -> io::Result<()> {
         }
     }
     Ok(())
+}
+
+#[derive(Default)]
+struct BrainfuckHighlighter;
+
+impl Highlighter for BrainfuckHighlighter {
+    fn highlight(&self, line: &str, _cursor: usize) -> StyledText {
+        let bf_style = Style::new().fg(Color::Cyan).bold();
+        let non_style = Style::new().fg(Color::DarkGray);
+        
+        let mut out: StyledText = StyledText::new();
+        
+        // Group contiguous BF and non-BF characters for styling
+        let mut current_is_bf: Option<bool> = None;
+        let mut current_buf = String::new();
+        
+        for ch in line.chars() {
+            let is_bf = matches!(ch, '>' | '<' | '+' | '-' | '.' | ',' | '[' | ']');
+            
+            match current_is_bf {
+                None => {
+                    current_is_bf = Some(is_bf);
+                    current_buf.push(ch);
+                }
+                Some(cur) if cur == is_bf => {
+                    current_buf.push(ch);
+                }
+                Some(_) => {
+                    // Flush previous span
+                    let style = if current_is_bf.unwrap() { bf_style } else { non_style };
+                    out.push((style, std::mem::take(&mut current_buf)));
+                    current_is_bf = Some(is_bf);
+                    current_buf.push(ch);
+                }
+            }
+        }
+        
+        if !current_buf.is_empty() {
+            let style = if current_is_bf.unwrap_or(false) { bf_style } else { non_style };
+            out.push((style, current_buf));
+        }
+        
+        out
+    }
 }
 
 #[cfg(test)]
